genbindings: malloc takes a size_t instead of a ulong

+++ cmd/handbindings/binding.go
@@ -16,7 +16,7 @@ import (
 
 func CArray(data []string) (C.int, **C.char) {
 
-	c_argv := (*[0xfff]*C.char)(C.malloc(C.ulong(8 /* sizeof pointer */ * len(data))))
+	c_argv := (*[0xfff]*C.char)(C.malloc(C.size_t(8 /* sizeof pointer */ * len(data))))
 
 	for i, arg := range data {
 		c_argv[i] = C.CString(arg)
